refactor: refactor the create product screen to sub components to resize the component line number
