+++ eccorce_app/app/src/main/java/com/example/eccomerce_app/ui/component/ProductVariantComponent.kt
@@ -1,6 +1,5 @@
 package com.example.eccomerce_app.ui.component
 
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -12,8 +11,12 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Clear
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -22,14 +25,17 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.constraintlayout.compose.ConstraintLayout
 import com.example.e_commercompose.R
 import com.example.e_commercompose.model.ProductVariant
+import com.example.e_commercompose.model.ProductVarientSelection
+import com.example.e_commercompose.model.VarirntModel
 import com.example.e_commercompose.ui.component.Sizer
 import com.example.e_commercompose.ui.theme.CustomColor
 import com.example.eccomerce_app.util.General
-import java.util.UUID
 
 
 @Composable
@@ -150,6 +156,78 @@ fun ProductVariantComponent(
                             }
                         }
                     }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun ProductVariantCreateComponent(
+    productVariants: List<ProductVarientSelection>,
+    variants: List<VarirntModel>? =null,
+    onRemoveProductVariant: (value: ProductVarientSelection) -> Unit
+){
+    FlowRow {
+        productVariants.forEach { value ->
+            ConstraintLayout(
+                modifier = Modifier.padding(end = 5.dp, bottom = 10.dp)
+            )
+            {
+                val (iconRef) = createRefs()
+                Column(
+                    modifier = Modifier
+                        .background(
+                            CustomColor.alertColor_3_300,
+                            RoundedCornerShape(8.dp)
+                        )
+                        .padding(
+                            end = 25.dp,
+                            start = 5.dp
+                        )
+                ) {
+                    Text(
+                        variants?.firstOrNull { it.id == value.variantId }?.name
+                            ?: "",
+                        fontFamily = General.satoshiFamily,
+                        fontWeight = FontWeight.Bold,
+                        fontSize = (18).sp,
+                        color = Color.White,
+                        textAlign = TextAlign.Center,
+                    )
+                    Text(
+                        value.name,
+                        fontFamily = General.satoshiFamily,
+                        fontWeight = FontWeight.Bold,
+                        fontSize = (18).sp,
+                        color = Color.White,
+                        textAlign = TextAlign.Center,
+                    )
+                }
+
+                Box(
+                    modifier = Modifier
+                        .height(20.dp)
+                        .width(20.dp)
+                        .background(
+                            Color.Red,
+                            RoundedCornerShape(20.dp)
+                        )
+                        .clip(
+                            RoundedCornerShape(20.dp)
+                        )
+                        .clickable { onRemoveProductVariant.invoke(value) }
+                        .constrainAs(iconRef) {
+                            top.linkTo(parent.top)
+                            end.linkTo(parent.end)
+                        },
+                    contentAlignment = Alignment.Center
+                ) {
+                    Icon(
+                        Icons.Default.Clear, "",
+                        tint = Color.White,
+                        modifier = Modifier.size(13.dp)
+                    )
                 }
             }
         }
