+++ eccorce_app/app/src/main/java/com/example/eccomerce_app/ui/view/account/store/CreateProduct.kt
@@ -1,62 +1,34 @@
 package com.example.eccomerce_app.ui.view.account.store
 
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
-import androidx.compose.material.icons.filled.Clear
-import androidx.compose.material.icons.filled.KeyboardArrowDown
 import androidx.compose.material3.BottomAppBar
-import androidx.compose.material3.CenterAlignedTopAppBar
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
@@ -64,19 +36,19 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.navigation.NavHostController
-import coil.compose.SubcomposeAsyncImage
 import com.example.e_commercompose.R
 import com.example.eccomerce_app.util.General
-import com.example.eccomerce_app.util.General.toCustomFil
 import com.example.e_commercompose.dto.ModelToDto.toListOfProductVarient
 import com.example.e_commercompose.model.ProductVarientSelection
 import com.example.e_commercompose.ui.component.CustomButton
+import com.example.e_commercompose.ui.component.CustomDropDownComponent
 import com.example.e_commercompose.ui.component.Sizer
 import com.example.eccomerce_app.ui.component.TextInputWithTitle
 import com.example.eccomerce_app.ui.component.TextNumberInputWithTitle
 import com.example.e_commercompose.ui.theme.CustomColor
+import com.example.eccomerce_app.ui.component.CreateProductImage
+import com.example.eccomerce_app.ui.component.ProductVariantCreateComponent
 import com.example.eccomerce_app.ui.component.SharedAppBar
 import com.example.eccomerce_app.viewModel.CurrencyViewModel
 import com.example.eccomerce_app.viewModel.ProductViewModel
@@ -126,14 +98,12 @@ fun CreateProductScreen(
 
     val images = remember { mutableStateOf(productData?.productImages ?: emptyList()) }
     val productVariants = remember {
-        mutableStateOf(
-            if (productData != null && !productData.productVariants.isNullOrEmpty()) productData.productVariants.toListOfProductVarient()
+        mutableStateOf(if (productData != null && !productData.productVariants.isNullOrEmpty()) productData.productVariants.toListOfProductVarient()
             else emptyList()
         )
     }
     val deleteImages = remember { mutableStateOf<List<String>>(emptyList()) }
-    val deleteProductVariant =
-        remember { mutableStateOf<List<ProductVarientSelection>>(emptyList()) }
+    val deleteProductVariant = remember { mutableStateOf<List<ProductVarientSelection>>(emptyList()) }
 
 
     val productName = remember { mutableStateOf(TextFieldValue("")) }
@@ -150,84 +120,14 @@ fun CreateProductScreen(
     val selectedVariantId = remember { mutableStateOf<UUID?>(null) }
 
 
-    val isExpandedSubCategory = remember { mutableStateOf(false) }
-    val isExpandedCurrency = remember { mutableStateOf(false) }
-    val isExpandedVariant = remember { mutableStateOf(false) }
     val isSendingData = remember { mutableStateOf(false) }
 
-
-    val animated = animateDpAsState(
-        if (isExpandedSubCategory.value) ((storeSubCategory?.size ?: 1) * 45).dp else 0.dp
-    )
-
-    val currencyAnimated = animateDpAsState(
-        if (isExpandedCurrency.value) ((currencies.value?.size ?: 1) * 45).dp else 0.dp
-    )
-    val rotation = animateFloatAsState(if (isExpandedSubCategory.value) 180f else 0f)
-    val currencyRotation = animateFloatAsState(if (isExpandedCurrency.value) 180f else 0f)
-    val animatedVariant = animateDpAsState(
-        if (isExpandedVariant.value) ((variants.value?.size ?: 1) * 45).dp else 0.dp
-    )
-    val rotationVariant = animateFloatAsState(if (isExpandedVariant.value) 180f else 0f)
 
     val snackBarHostState = remember { SnackbarHostState() }
 
 
-    val onImageSelection = rememberLauncherForActivityResult(
-        ActivityResultContracts.PickVisualMedia()
-    )
-    { uri ->
-        if (uri != null) {
-            val fileHolder = uri.toCustomFil(context = context)
-            if (fileHolder != null) {
-                if (thumbnail.value != null && !deleteImages.value.contains(thumbnail.value)) {
-                    val deleteImageCopy = mutableListOf<String>()
-                    deleteImageCopy.add(thumbnail.value!!)
-                    deleteImageCopy.addAll(deleteImages.value)
-                    deleteImages.value = deleteImageCopy
-                }
-                thumbnail.value = fileHolder.absolutePath
-
-            }
-        }
-    }
-
-
-    val selectMultipleImages = rememberLauncherForActivityResult(
-        ActivityResultContracts.PickMultipleVisualMedia(10)
-    )
-    { uris ->
-        val imagesHolder = mutableListOf<String>()
-
-        if (uris.isNotEmpty()) {
-            uris.forEach { productImages ->
-                val file = productImages.toCustomFil(context)
-
-                if (file != null) {
-                    imagesHolder.add(file.absolutePath)
-                }
-            }
-            if (imagesHolder.isNotEmpty()) {
-                imagesHolder.addAll(images.value)
-                images.value = imagesHolder
-            }
-        }
-    }
-
-    fun updateConditionValue(
-        isExpandedSubCategoryValue: Boolean? = null,
-        isExpandedCurrencyValue: Boolean? = null,
-        isExpandedVariantValue: Boolean? = null,
-        isSendingDataValue: Boolean? = null
-    ) {
-        when {
-            isExpandedSubCategoryValue != null -> isExpandedSubCategory.value =
-                isExpandedSubCategoryValue
-
-            isExpandedCurrencyValue != null -> isExpandedCurrency.value = isExpandedCurrencyValue
-            isExpandedVariantValue != null -> isExpandedVariant.value = isExpandedVariantValue
-            isSendingDataValue != null -> isSendingData.value = isSendingDataValue
-        }
+    fun updateConditionValue(isSendingDataValue: Boolean) {
+          isSendingData.value = isSendingDataValue
     }
 
     fun validateInput(): Boolean {
@@ -256,22 +156,20 @@ fun CreateProductScreen(
     }
 
     //this for api product images not for create this only when product is update
-    fun removeProductImages(image: String) {
-        if (productData != null && productData.productImages.contains(image)
-        ) {
-            val deleteImageList = mutableListOf<String>()
-            deleteImageList.add(image)
-            deleteImageList.addAll(deleteImages.value)
-            deleteImages.value = deleteImageList
+    fun removeProductImages(deletedImage: List<String>) {
+        if(thumbnail.value==deletedImage[0]){
+            deleteImages.value +=deletedImage[0]
         }
-        images.value = images.value.filter { it -> it != image }
+        else {
+            val removedDeletedImages = images.value.filter { !deletedImage.contains(it) }
+            images.value = removedDeletedImages
+        }
     }
 
     //this for api product product variant not for create this only when product is update
     fun removeProductVariant(value: ProductVarientSelection) {
         if (productData != null && !productData.productVariants.isNullOrEmpty() &&
-            productData.productVariants.toListOfProductVarient()
-                .contains(value)
+            productData.productVariants.toListOfProductVarient().firstOrNull{it==value}!=null
         ) {
             val deletedProductVariant =
                 mutableListOf<ProductVarientSelection>()
@@ -285,18 +183,11 @@ fun CreateProductScreen(
 
     fun updateCurrentCurrencySymbol(currencySymbol: String) {
         productCurrency.value = currencySymbol
-        updateConditionValue(isExpandedCurrencyValue = false)
     }
 
-    fun updateSubCategory(id: UUID) {
-        updateConditionValue(isExpandedSubCategoryValue = false)
-        selectedSubCategoryId.value = id
-    }
+    fun updateSubCategory(id: UUID) { selectedSubCategoryId.value = id }
 
-    fun updateSelectedVariant(id: UUID) {
-        updateConditionValue(isExpandedVariantValue = false)
-        selectedVariantId.value = id
-    }
+    fun updateSelectedVariant(id: UUID) { selectedVariantId.value = id }
 
     fun addProductVariant() {
         val selectedVariant = ProductVarientSelection(
@@ -468,7 +359,8 @@ fun CreateProductScreen(
                 )
             }
         }
-    ) { scaffoldStatus ->
+    )
+    { scaffoldStatus ->
         scaffoldStatus.calculateTopPadding()
         scaffoldStatus.calculateBottomPadding()
 
@@ -478,271 +370,19 @@ fun CreateProductScreen(
                 .background(Color.White)
                 .padding(scaffoldStatus)
                 .padding(horizontal = 15.dp)
-                .padding(top = scaffoldStatus.calculateTopPadding() + 30.dp)
                 .verticalScroll(rememberScrollState())
         )
         {
 
-            Text(
-                stringResource(R.string.product_thumbnail),
-                fontFamily = General.satoshiFamily,
-                fontWeight = FontWeight.Bold,
-                fontSize = (18).sp,
-                color = CustomColor.neutralColor950,
-                textAlign = TextAlign.Center,
+            CreateProductImage(
+                thumbnail = thumbnail.value,
+                images = images.value,
+                context = context,
+                deleteImages = deleteImages.value,
+                onSelectThumbnail = {value->thumbnail.value=value},
+                onSelectImages = {value-> images.value += value },
+                onRemoveAlreadyProductImage = {value-> removeProductImages(value)}
             )
-            Sizer(15)
-            ConstraintLayout(
-                modifier = Modifier
-                    .fillMaxWidth()
-            )
-            {
-                val (imageRef, cameralRef) = createRefs()
-                Box(
-                    modifier = Modifier
-                        .constrainAs(imageRef) {
-                            top.linkTo(parent.top)
-                            bottom.linkTo(parent.bottom)
-                            start.linkTo(parent.start)
-                            end.linkTo(parent.end)
-                        }
-                        .height(150.dp)
-                        .fillMaxWidth()
-                        .border(
-                            width = 1.dp,
-                            color = CustomColor.neutralColor500,
-                            shape = RoundedCornerShape(8.dp)
-                        )
-                        .background(
-                            color = Color.White,
-                        ),
-                    contentAlignment = Alignment.Center
-                )
-                {
-                    when (thumbnail.value == null) {
-                        true -> {
-                            Icon(
-                                imageVector = ImageVector.vectorResource(R.drawable.insert_photo),
-                                "",
-                                modifier = Modifier.size(80.dp),
-                                tint = CustomColor.neutralColor200
-                            )
-                        }
-
-                        else -> {
-                            SubcomposeAsyncImage(
-                                contentScale = ContentScale.Crop,
-                                modifier = Modifier
-                                    .fillMaxHeight()
-                                    .fillMaxWidth()
-                                    .clip(RoundedCornerShape(8.dp)),
-                                model = General.handlingImageForCoil(
-                                    thumbnail.value,
-                                    context
-                                ),
-                                contentDescription = "",
-                                loading = {
-                                    Box(
-                                        modifier = Modifier
-                                            .fillMaxSize(),
-                                        contentAlignment = Alignment.Center // Ensures the loader is centered and doesn't expand
-                                    ) {
-                                        CircularProgressIndicator(
-                                            color = Color.Black,
-                                            modifier = Modifier.size(54.dp) // Adjust the size here
-                                        )
-                                    }
-                                },
-                            )
-                        }
-                    }
-
-                }
-                Box(
-                    modifier = Modifier
-                        .padding(end = 5.dp, bottom = 10.dp)
-                        .constrainAs(cameralRef) {
-                            end.linkTo(imageRef.end)
-                            bottom.linkTo(imageRef.bottom)
-                        }
-
-
-                )
-                {
-
-                    IconButton(
-                        onClick = {
-                            onImageSelection.launch(
-                                PickVisualMediaRequest(
-                                    ActivityResultContracts.PickVisualMedia.ImageOnly
-                                )
-                            )
-                        },
-                        modifier = Modifier
-                            .size(30.dp),
-                        colors = IconButtonDefaults.iconButtonColors(
-                            containerColor = CustomColor.primaryColor200
-                        )
-                    ) {
-                        Icon(
-                            ImageVector.vectorResource(R.drawable.camera),
-                            "",
-                            modifier = Modifier.size(18.dp),
-                            tint = Color.White
-                        )
-                    }
-                }
-
-            }
-            Sizer(10)
-            Text(
-                stringResource(R.string.product_images),
-                fontFamily = General.satoshiFamily,
-                fontWeight = FontWeight.Bold,
-                fontSize = (18).sp,
-                color = CustomColor.neutralColor950,
-                textAlign = TextAlign.Center,
-            )
-            Sizer(5)
-            ConstraintLayout(
-                modifier = Modifier
-                    .fillMaxWidth()
-            )
-            {
-                val (cameralRef) = createRefs()
-
-                FlowRow(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .border(
-                            width = 1.dp,
-                            color = CustomColor.neutralColor500,
-                            shape = RoundedCornerShape(8.dp)
-                        )
-                        .padding(horizontal = 5.dp, vertical = 5.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalArrangement = Arrangement.spacedBy(5.dp)
-
-                ) {
-                    if (images.value.isEmpty()) {
-                        Box(
-                            modifier = Modifier
-                                .height(150.dp)
-                                .fillMaxWidth(),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Icon(
-                                imageVector = ImageVector.vectorResource(R.drawable.insert_photo),
-                                "",
-                                modifier = Modifier.size(80.dp),
-                                tint = CustomColor.neutralColor200
-                            )
-                        }
-                    }
-
-                    images.value.forEach { value ->
-
-                        ConstraintLayout {
-                            Box(
-                                modifier = Modifier
-
-                                    .height(120.dp)
-                                    .width(120.dp)
-                                    .background(
-                                        color = Color.White,
-                                    ),
-                                contentAlignment = Alignment.Center
-                            ) {
-                                SubcomposeAsyncImage(
-                                    contentScale = ContentScale.Crop,
-                                    modifier = Modifier
-                                        .fillMaxHeight()
-                                        .fillMaxWidth()
-                                        .clip(RoundedCornerShape(8.dp)),
-                                    model = General.handlingImageForCoil(
-                                        value,
-                                        context
-                                    ),
-                                    contentDescription = "",
-                                    loading = {
-                                        Box(
-                                            modifier = Modifier
-                                                .fillMaxSize(),
-                                            contentAlignment = Alignment.Center // Ensures the loader is centered and doesn't expand
-                                        ) {
-                                            CircularProgressIndicator(
-                                                color = Color.Black,
-                                                modifier = Modifier.size(54.dp) // Adjust the size here
-                                            )
-                                        }
-                                    },
-                                )
-
-                            }
-
-                            Box(
-                                modifier = Modifier
-                                    .height(30.dp)
-                                    .width(30.dp)
-                                    .background(
-                                        Color.Red,
-                                        RoundedCornerShape(20.dp)
-                                    )
-                                    .clip(
-                                        RoundedCornerShape(20.dp)
-                                    )
-                                    .clickable {
-                                        removeProductImages(value)
-                                    },
-                                contentAlignment = Alignment.Center
-                            )
-                            {
-                                Icon(
-                                    Icons.Default.Clear, "",
-                                    tint = Color.White
-                                )
-                            }
-                        }
-
-                    }
-                }
-
-
-                Box(
-                    modifier = Modifier
-                        .padding(end = 5.dp, bottom = 10.dp)
-                        .constrainAs(cameralRef) {
-                            end.linkTo(parent.end)
-                            bottom.linkTo(parent.bottom)
-                        }
-                )
-                {
-
-                    IconButton(
-                        onClick = {
-                            selectMultipleImages.launch(
-                                PickVisualMediaRequest(
-                                    ActivityResultContracts.PickVisualMedia.ImageOnly
-                                )
-                            )
-                        },
-                        modifier = Modifier
-                            .size(30.dp),
-                        colors = IconButtonDefaults.iconButtonColors(
-                            containerColor = CustomColor.primaryColor200
-                        )
-                    ) {
-                        Icon(
-                            ImageVector.vectorResource(R.drawable.camera),
-                            "",
-                            modifier = Modifier.size(18.dp),
-                            tint = Color.White
-                        )
-                    }
-                }
-
-            }
-
             Sizer(15)
 
             TextInputWithTitle(
@@ -780,74 +420,13 @@ fun CreateProductScreen(
                 textAlign = TextAlign.Center,
             )
             Sizer(5)
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .border(
-                        1.dp,
-                        CustomColor.neutralColor400,
-                        RoundedCornerShape(8.dp)
-                    )
-                    .clip(RoundedCornerShape(8.dp))
+            CustomDropDownComponent(
+                value =  productData?.symbol ?: (productCurrency.value ?: "Chose Currency"),
+                items =  currencies.value?.map { it.name }?:emptyList(),
+                onSelectValue = { value ->
+                    updateCurrentCurrencySymbol(value)
+                }
             )
-            {
-                Row(
-                    modifier = Modifier
-                        .height(65.dp)
-                        .fillMaxWidth()
-
-                        .clickable {
-                            updateConditionValue(isExpandedCurrencyValue = !isExpandedCurrency.value)
-                        }
-                        .padding(horizontal = 5.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                )
-                {
-                    Text(
-                        productData?.symbol ?: (productCurrency.value ?: "Chose Currency")
-                    )
-                    Icon(
-                        Icons.Default.KeyboardArrowDown, "",
-                        modifier = Modifier.rotate(currencyRotation.value)
-                    )
-                }
-
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(currencyAnimated.value)
-                        .border(
-                            1.dp,
-                            CustomColor.neutralColor200,
-                            RoundedCornerShape(
-                                topStart = 0.dp,
-                                topEnd = 0.dp,
-                                bottomStart = 8.dp,
-                                bottomEnd = 8.dp
-                            )
-                        ),
-
-                    )
-                {
-                    currencies.value?.forEach { value ->
-                        Text(
-                            value.name,
-                            modifier = Modifier
-                                .height(50.dp)
-                                .fillMaxWidth()
-                                .clip(RoundedCornerShape(8.dp))
-                                .clickable {
-                                    updateCurrentCurrencySymbol(value.symbol)
-                                }
-                                .padding(top = 12.dp, start = 5.dp)
-
-                        )
-                    }
-                }
-            }
-
-
             Sizer(15)
 
             Text(
@@ -859,78 +438,19 @@ fun CreateProductScreen(
                 textAlign = TextAlign.Center,
             )
             Sizer(5)
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .border(
-                        1.dp,
-                        CustomColor.neutralColor400,
-                        RoundedCornerShape(8.dp)
-                    )
-                    .clip(RoundedCornerShape(8.dp))
+            CustomDropDownComponent(
+               value = if (productData != null && selectedSubCategoryId.value == null)
+                   storeSubCategory?.firstOrNull { it.id == productData.subcategoryId }?.name
+                       ?: stringResource(R.string.select_subcategory)
+               else if (selectedSubCategoryId.value == null) stringResource(R.string.select_subcategory)
+               else storeSubCategory?.firstOrNull { it.id == selectedSubCategoryId.value }?.name
+                   ?: "",
+                items = storeSubCategory?.map { it.name },
+                onSelectValue = { value ->
+                    updateSubCategory(storeSubCategory?.firstOrNull { it.name == value }?.id ?: UUID.randomUUID())
+                }
             )
-            {
 
-                Row(
-                    modifier = Modifier
-                        .height(65.dp)
-                        .fillMaxWidth()
-
-                        .clickable {
-                            updateConditionValue(isExpandedSubCategoryValue = !isExpandedSubCategory.value)
-                        }
-                        .padding(horizontal = 5.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                )
-                {
-                    Text(
-                        if (productData != null && selectedSubCategoryId.value == null)
-                            storeSubCategory?.firstOrNull { it.id == productData.subcategoryId }?.name
-                                ?: stringResource(R.string.select_subcategory)
-                        else if (selectedSubCategoryId.value == null) stringResource(R.string.select_subcategory)
-                        else storeSubCategory?.firstOrNull { it.id == selectedSubCategoryId.value }?.name
-                            ?: ""
-                    )
-                    Icon(
-                        Icons.Default.KeyboardArrowDown, "",
-                        modifier = Modifier.rotate(rotation.value)
-                    )
-                }
-
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(animated.value)
-                        .border(
-                            1.dp,
-                            CustomColor.neutralColor200,
-                            RoundedCornerShape(
-                                topStart = 0.dp,
-                                topEnd = 0.dp,
-                                bottomStart = 8.dp,
-                                bottomEnd = 8.dp
-                            )
-                        ),
-
-                    )
-                {
-                    storeSubCategory?.forEach { value ->
-                        Text(
-                            value.name,
-                            modifier = Modifier
-                                .height(50.dp)
-                                .fillMaxWidth()
-                                .clip(RoundedCornerShape(8.dp))
-                                .clickable {
-                                    updateSubCategory(value.id)
-                                }
-                                .padding(top = 12.dp, start = 5.dp)
-
-                        )
-                    }
-                }
-            }
 
             Sizer(5)
 
@@ -947,139 +467,25 @@ fun CreateProductScreen(
                 Sizer(2)
                 if (productVariants.value.isNotEmpty())
                     Sizer(5)
-                FlowRow {
-                    productVariants.value.forEach { value ->
-                        ConstraintLayout(
-                            modifier = Modifier.padding(end = 5.dp, bottom = 10.dp)
-                        ) {
-                            val (iconRef) = createRefs()
-                            Column(
-                                modifier = Modifier
-                                    .background(
-                                        CustomColor.alertColor_3_300,
-                                        RoundedCornerShape(8.dp)
-                                    )
-                                    .padding(
-                                        end = 25.dp,
-                                        start = 5.dp
-                                    )
-                            ) {
-                                Text(
-                                    variants.value?.firstOrNull { it.id == value.variantId }?.name
-                                        ?: "",
-                                    fontFamily = General.satoshiFamily,
-                                    fontWeight = FontWeight.Bold,
-                                    fontSize = (18).sp,
-                                    color = Color.White,
-                                    textAlign = TextAlign.Center,
-                                )
-                                Text(
-                                    value.name,
-                                    fontFamily = General.satoshiFamily,
-                                    fontWeight = FontWeight.Bold,
-                                    fontSize = (18).sp,
-                                    color = Color.White,
-                                    textAlign = TextAlign.Center,
-                                )
-                            }
 
-                            Box(
-                                modifier = Modifier
-                                    .height(20.dp)
-                                    .width(20.dp)
-                                    .background(
-                                        Color.Red,
-                                        RoundedCornerShape(20.dp)
-                                    )
-                                    .clip(
-                                        RoundedCornerShape(20.dp)
-                                    )
-                                    .clickable { removeProductVariant(value) }
-                                    .constrainAs(iconRef) {
-                                        top.linkTo(parent.top)
-                                        end.linkTo(parent.end)
-                                    },
-                                contentAlignment = Alignment.Center
-                            ) {
-                                Icon(
-                                    Icons.Default.Clear, "",
-                                    tint = Color.White,
-                                    modifier = Modifier.size(13.dp)
-                                )
-                            }
-                        }
-                    }
-                }
-                if (productVariants.value.isNotEmpty())
-                    Sizer(5)
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .border(
-                            1.dp,
-                            CustomColor.neutralColor400,
-                            RoundedCornerShape(8.dp)
-                        )
-                        .clip(RoundedCornerShape(8.dp))
+                ProductVariantCreateComponent(
+                   productVariants =  productVariants.value,
+                   variants = variants.value,
+                onRemoveProductVariant =  { value->removeProductVariant(value)}
                 )
-                {
+                 if (productVariants.value.isNotEmpty())
+                    Sizer(5)
 
-                    Row(
-                        modifier = Modifier
-                            .height(65.dp)
-                            .fillMaxWidth()
-                            .clickable {
-                                updateConditionValue(isExpandedVariantValue = !isExpandedVariant.value)
-                            }
-                            .padding(horizontal = 5.dp),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    )
-                    {
-                        Text(
-                            if (selectedVariantId.value == null) stringResource(R.string.select_variant)
-                            else variants.value?.firstOrNull { it.id == selectedVariantId.value }?.name
-                                ?: ""
-                        )
-                        Icon(
-                            Icons.Default.KeyboardArrowDown, "",
-                            modifier = Modifier.rotate(rotationVariant.value)
-                        )
+                CustomDropDownComponent(
+                    value = if (selectedVariantId.value == null) stringResource(R.string.select_variant)
+                    else variants.value?.firstOrNull { it.id == selectedVariantId.value }?.name
+                        ?: "",
+                    items = variants.value?.map { it.name },
+                    onSelectValue = { value ->
+                        updateSelectedVariant(variants.value?.firstOrNull { it.name == value }?.id ?: UUID.randomUUID())
                     }
+                )
 
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(animatedVariant.value)
-                            .border(
-                                1.dp,
-                                CustomColor.neutralColor200,
-                                RoundedCornerShape(
-                                    topStart = 4.dp,
-                                    topEnd = 4.dp,
-                                    bottomStart = 8.dp,
-                                    bottomEnd = 8.dp
-                                )
-                            ),
-
-                        ) {
-                        if (!variants.value.isNullOrEmpty())
-                            variants.value!!.forEach { value ->
-                                Text(
-                                    value.name,
-                                    modifier = Modifier
-                                        .height(50.dp)
-                                        .fillMaxWidth()
-                                        .clip(RoundedCornerShape(8.dp))
-                                        .clickable {
-                                            updateSelectedVariant(value.id)
-                                        }
-                                        .padding(top = 12.dp, start = 5.dp)
-
-                                )
-                            }
-                    }
-                }
                 Sizer(5)
                 OutlinedTextField(
                     maxLines = 6,
